requirements.txt
openai>=1.3.0
requests>=2.31.0
python-dotenv>=1.0.0
Pillow>=10.0.0
moviepy>=1.0.3
schedule>=1.2.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
tenacity>=8.2.0